% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownian_motion.R
\name{brownian.motion}
\alias{brownian.motion}
\title{Simulate paths of the standard Brownian Motion on [0,1].}
\usage{
brownian.motion(n, p)
}
\arguments{
\item{n}{Number of paths to simulate.}

\item{p}{Number of discretisation points (exluding 0) for each path.}
}
\value{
A matrix (p+1 by n) of simulated independent paths of the standard Brownian motion on [0,1].
}
\description{
Simulate paths of the standard Brownian Motion on [0,1].
}
\examples{
x <- brownian.motion(20,200)
matplot(seq(0,1,1/200),x, type="l", xlab="t", ylab="x(t)")
}
