% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_covariance.R
\name{compute_covariance}
\alias{compute_covariance}
\title{Compute the conditional covariance matrix given X.}
\usage{
compute_covariance(p, neighbor.indices)
}
\arguments{
\item{p}{Vector (length K)}

\item{neighbor.indices}{Vector (length n)}
}
\value{
The conditional covariance matrix.
}
\description{
Computes the conditional covariance matrix given X of the empirical deviations from independence under the null hypothesis.
Only intended for use in the \link{psicor.test} funciton.
}
\examples{
library(RANN)
x <- brownian.motion(100, 200)
p <- c(0.3, 0.7)

neighbor.indices <- nn2(x, k=2)$nn.idx[, 2]
sigma <- compute_covariance(p, neighbor.indices)
}
\references{
Hörmann, S. and Strenger, D. (2025). Azadkia–Chatterjee’s dependence coefficient for infinite dimensional data.
}
\author{
Daniel Strenger
}
