% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psicor.R
\name{psicor}
\alias{psicor}
\title{Estimate the psi correlation.}
\usage{
psicor(Y, Z, X = NULL, eps = 2, na.rm = TRUE)
}
\arguments{
\item{Y}{Vector (length n)}

\item{Z}{Matrix (n by p)}

\item{X}{Matrix (n by q), default is NULL}

\item{eps}{Error bound for nearest neighbour search: a value of 0.0 implies exact nearest neighbour search.}

\item{na.rm}{Remove NAs if TRUE}
}
\value{
Psi correlation of Y and Z given X. If X == NULL, this is just a measure of the dependence between Y and Z.
}
\description{
Psi correlation is a measure of the amount of conditional dependence between a categorical random variable Y and a random variable Z taking values in a separable metric space given a random variable X taking values in a separable metric space, based on an i.i.d. sample of (Y,Z,X). The coefficient is asymptotically guaranteed to be between 0 and 1.
}
\details{
The value returned by psicor can be positive or negative. Asymptotically, it is guaranteed to be between 0 and 1. If X==NULL, the finite sample version is non-negative as well. psicor is invariant under permutation of the classes of Y.
}
\examples{
Z <- matrix(runif(100*2), ncol=2)
Y <- sin(2*pi*rowSums(Z)) >= 0 # A highly non-linear relationship
psicor(Y, Z)
}
\references{
Hörmann, S. and Strenger, D. (2025).  TITLE. \href{}{}
}
\author{
Daniel Strenger
}
