% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psicor.test.R
\name{psicor.test}
\alias{psicor.test}
\title{Psi correlation test for categorical variables.}
\usage{
psicor.test(Y, X, eps = 2, na.rm = TRUE, permutation = FALSE, R = 1000)
}
\arguments{
\item{Y}{Vector (length n) of categorical values}

\item{X}{Matrix (n by q)}

\item{eps}{Error bound for nearest neighbour search: a value of 0.0 implies exact nearest neighbour search.}

\item{na.rm}{Remove NAs if TRUE}

\item{permutation}{If TRUE, the p-value is calculated by permutation. If FALSE, the p-value is calculated based on the asymptotic distribution of the test statistic. The default is FALSE.}

\item{R}{Number of permutations. Ignored if permutation == FALSE.}
}
\value{
A list containing the test statistic and its p-value.
}
\description{
Psi correlation test for categorical variables.
}
\details{
The test statistic is a mean square contingency, similar to the usual chi-squared divergence.
It is known to be asymptotically distributed as a chi-squared distribution with K*(K-1) degrees of freedom, where K is the number of levels of Y. The p-value is either calculated based on this distribution or on a resampling procedure, according to the value of the argument \code{permutation}.
}
\examples{
X <- matrix(runif(100*2), ncol=2)
Y <- sin(2*pi*rowSums(X)) >= 0 # A highly non-linear relationship
psicor.test(Y,X)
}
\references{
Hörmann, S. and Strenger, D. (2025).  TITLE. \href{}{}
}
\author{
Daniel Strenger
}
