% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psicor_vs.R
\name{psicor.vs}
\alias{psicor.vs}
\title{Variable selection based on psi correlation.}
\usage{
psicor.vs(Y, X, threshold = 0, num_variables = Inf, cores = 1, verbose = TRUE)
}
\arguments{
\item{Y}{Vector (length n)}

\item{X}{Matrix (n by p)}

\item{threshold}{Threshold for psi correlation: the algorithm stops if the conditional psi correlation is below
this value.}

\item{num_variables}{Maximum number of features to select. If the number of features selected
reaches this value, the algorithm stops.}

\item{cores}{Number of cores to use for parallel computation. If cores > 1, the function uses parallel computation.}

\item{verbose}{If TRUE, the function prints progress messages}
}
\value{
Psi correlation of Y and Z given X. If X == NULL, this is just a measure of the dependence between Y and Z.
}
\description{
A variable selection algorithm on psi correlation \link{psicor}.
}
\details{
A forward stepwise variable selection algorithm using on psi correlation \link{psicor} at each step.
It terminates when the conditional psi correlation of any possible additional
variable and the response given the already selected variables is below \code{threshold}
or if \code{num_variables} variables are selected. To reduce the computation time,
in each step the conditional psi correlation is calculated for all remaining variables in parallel if \code{cores}>1.
}
\examples{
library(kernlab)
data(spam)
X <- spam[, -58]
Y <- spam[, 58]
psicor.vs(Y, X, threshold = 0, cores = parallel::detectCores())
}
\references{
Hörmann, S. and Strenger, D. (2025).  TITLE. \href{}{}
}
\author{
Daniel Strenger
}
