% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soFun.R
\name{soFun}
\alias{soFun}
\title{Estimate the Scalar on Function Dependence Coefficient (soFun).}
\usage{
soFun(Y, X, eps = 2, na.rm = TRUE)
}
\arguments{
\item{Y}{Vector of scalar responses (length n).}

\item{X}{Matrix of functional covariates (p by n).}

\item{eps}{Error bound for nearest neighbour search: a value of 0.0 implies exact nearest neighbour search.}

\item{na.rm}{Remove NAs if TRUE}
}
\value{
The measure of dependence of Y on X.
}
\description{
The Scalar on Function Dependence Coefficient (soFun) is an adaption of the conditional dependence coefficient (CODEC) by Azadkia and Chatterjee to the setting of continuously distributed response Y and functional covariate X. It is a measure of the amount of dependence of Y on X, based on an i.i.d. sample of (Y, X). The coefficient is asymptotically guaranteed to be between 0 and 1.
}
\details{
The returned statistic can be positive or negative. Asymptotically, it is guaranteed to be between 0 and 1. A small value indicates low dependence of Y on X, and a high value indicates strong dependence. The statistic is the base for the \code{\link{soFun.test}}.
}
\examples{
n <- 100
p <- 200
x <- brownian.motion(n,p)
y <- sin(2*pi* apply(x,2,mean))
soFun(y,x)
}
\references{
Hörmann, S. and Strenger, D. (2025). Azadkia–Chatterjee’s dependence coefficient for infinite dimensional data. \href{https://arxiv.org/abs/2405.07732}{https://arxiv.org/abs/2405.07732}

Azadkia, M. and Chatterjee, S. (2019). A simple measure of conditional dependence. \href{https://arxiv.org/pdf/1910.12327.pdf}{https://arxiv.org/pdf/1910.12327.pdf}.
}
