% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soFun.R
\name{soFun.test}
\alias{soFun.test}
\title{Scalar on Function test for dependence.}
\usage{
soFun.test(Y, X, eps = 2, na.rm = TRUE)
}
\arguments{
\item{Y}{Vector (length n)}

\item{X}{Matrix (p by n)}

\item{eps}{Error bound for nearest neighbour search: a value of 0.0 implies exact nearest neighbour search.}

\item{na.rm}{Remove NAs if TRUE}
}
\value{
A list containing the test statistic and its p-value.
}
\description{
The test is based on the asymptotic distribution of the soFun statistic.
}
\details{
The test statistic is an appropriately scaled version of the \code{\link{soFun}} statistic.
The test is based on the known asymptotic normal distribution of this statistic, not on a resampling procedure.
}
\examples{
n <- 100
p <- 200
x <- brownian.motion(n,p)
y <- sin(2*pi* apply(x,2,mean))
if(soFun.test(y,x)$p.value < 0.05){
  print("The hypothesis of independence can be rejected at a significance level of 0.05.")
} else{
  print("The hypothesis of independence cannot be rejected at a significance level of 0.05.")
}
}
\references{
Hörmann, S. and Strenger, D. (2024). Measuring dependence between a scalar response and a functional covariate.
}
\author{
Daniel Strenger
}
